chore: separate time related functions to a seperate file
