+++ src/pages/watch/[id].tsx
@@ -14,6 +14,7 @@ import Episode from '@components/watch/Episode';
 import { AnimeBannerFragment, AnimeInfoFragment } from '@generated/aniList';
 import useAnime from '@hooks/useAnime';
 import { watchPage } from '@lib/api';
+import { convertToDate, convertToTime } from '@utility/time';
 import { proxyUrl } from '@utility/utils';
 
 const VideoPlayer = dynamic(() => import('@components/watch/VideoPlayer'), {
@@ -49,7 +50,7 @@ export const getServerSideProps: GetServerSideProps<WatchProps> = async (
   };
 };
 
-const Video = ({
+const Watch = ({
   anime,
   recommended,
 }: InferGetServerSidePropsType<typeof getServerSideProps>) => {
@@ -134,53 +135,6 @@ const Video = ({
     }
     setProxy(!videoLink.match(urls));
   }, [videoLink, urls]);
-
-  const nth = (days: number) => {
-    if (days > 3 && days < 21) return 'th';
-
-    switch (days % 10) {
-      case 1:
-        return 'st';
-      case 2:
-        return 'nd';
-      case 3:
-        return 'rd';
-      default:
-        return 'th';
-    }
-  };
-
-  const convertToDate = (date: number) => {
-    const dateObj = new Date(date);
-    const year = dateObj.getFullYear();
-    const month = dateObj.toLocaleString('en-us', { month: 'short' });
-    const day = dateObj.getDate();
-
-    return `${day}${nth(day)} of ${month} ${year}`;
-  };
-
-  const convertToTime = (date: number) => {
-    const weekday = [
-      'Sunday',
-      'Monday',
-      'Tuesday',
-      'Wednesday',
-      'Thursday',
-      'Friday',
-      'Saturday',
-    ];
-
-    const dateObj = new Date(date);
-    const day = weekday[dateObj.getDay()];
-    let hours = dateObj.getHours();
-    let minutes: number | string = dateObj.getMinutes();
-    const ampm = hours >= 12 ? 'pm' : 'am';
-    hours %= 12;
-    hours = hours || 12; // the hour '0' should be '12'
-    minutes = minutes < 10 ? `0${minutes}` : minutes;
-    const strTime = `${day} at ${hours}:${minutes} ${ampm} (GMT-4)`;
-    return strTime;
-  };
 
   return (
     <>
@@ -300,4 +254,4 @@ const Video = ({
   );
 };
 
-export default Video;
+export default Watch;
